gettext.c: treat GIT_USE_PREFERENCES_LOCALE as a boolean
Use git_env_bool on GIT_USE_PREFERENCES_LOCALE so that
the value can be used to disable as well as enable the option.

Unfortunately this means if GIT_USE_PREFERENCES_LOCALE is
set to an invalid value (such as "a") then git will exit
immediately with a fatal (untranslated) error message.