add WWW::Curl perl module (patched) to the build
git-send-email.perl needs to be changed to use WWW::Curl to send
smtp and smtps email.  In order for that to work it needs a patched
version of WWW::Curl that can handle the CURLOPT_MAIL_RCPT option.

Since we are already building cURL and WWW::Curl has the same
license, no dependencies and is small, we just include it (with
the necessary patch) too.